Added code to the start-up class (Program.cs) to only allow a single instance of Cryptnos to be running at a time. Running more than one copy at once could corrupt the site list in the registry. No one has reported a problem with this, but I know I encountered the situation once during testing and didn't want it to happen to any users.
Changed the length restriction (character limit) text box to a drop-down list, similar to the recent changes in Cryptnos for Android 1.2.1. The reason for the change is two-fold: (1) to reduce user confusion by having to leave the text box empty when no restriction is needed and (2) to limit the possible values the user can select. The drop-down dynamically changes its available values based on the user's selection of hash algorithm. If the user selects a new hash that has a maximum length shorter than the current limit, the limit is effectively changed to "None". Otherwise, the preference is retained regardless of the new hash maximum.

Added some stricter validation on the iterations text box, forcing it to explicitly deny any non-numeric input and any leading zeros. I considered making this a drop-down list or a "spinner" ("numeric up or down" in .NET parlance), but found that it wasn't as restrictive as I wanted it to be. The text box works just fine, and is even more restrictive thanks to the new validation code.

Also changed the version display string in the main form to include the build number. Previously, it only showed "major.minor"; now it shows "major.minor.build". Since I've been making a number of build-level revisions lately, it seemed silly not to include that. Revision is still unnecessary and is still only displayed in the About dialog.