Update Issue 9
A number of first attempts to improve support for Cryptnos under Mono.

MoMA reports one P/Invoke and one "to-do" warning, both of which are used to prevent multiple instances of the app from running at the same time under Windows. This code obviously doesn't work under other OSes, so we check to see if the app is running under Mono and, if so, we skip that check. It will be upon the user's shoulders to kill other open instances before a new instance can be launched. The caveat will be that we will get this same behavior for Windows users running under Mono, but since the code is .NET specific, we can't avoid that.

Several of the "dialog boxes" (actually other forms) use a little kludge for passing values back to the main form. This kludge causes a stack overflow under Mono. If we detect that we're running under Mono, we now avoid the kludge and let the flow go without it; it still seems to produce the desired behavior.

Our main form is *REALLY* poorly designed, and Mono highlights this. Specifically, switching to "daily mode" crops parts of the form (at least under GNOME/Linux) and renders it difficult if not impossible to use. For the time being, we are disabling "daily mode" when Cryptnos is run under Mono until we can improve this behavior.

The "copy password to clipboard" feature technically works, but perhaps not as expected. Under Linux, there are often two separate clipboards: one use by X applications and one used by command-line apps (Ctrl+C and Ctrl+V). Cryptnos successfully copies to one of these, but I'm not entirely sure which. Unfortunately, it may take some experimentation on the user's part to find out for themselves.

The URL link for the Cryptnos site in the About dialog does not seem to launch the browser, although the Help button does successfully launch the local help file. Go figure.

There are still a few visual artifacts under GNOME/Linux, but as far as I can tell, the app seems fully functional. There's still a need for extensive testing, of course.

I have done absolutely no testing under Mac OS X, since I unfortunately do not have a Mac to test with. We may have to claim OS X is officially unsupported until we can find a volunteer tester or I finally break down and buy a Mac. (I've been thinking of tinkering with iOS development anyway.)

Tentatively bumping this to version 1.3.4. (No tags just yet.)