Update Issue 9
Upgraded GPFUpdateChecker to version 1.2. Now when a new update is found, the behavior is different under Windows/.NET and any platform running Mono. Under Windows/.NET, the old behavior is maintained: if the user says it's OK, the installer will be downloaded and executed automatically. Under Mono, however, the user will be notified, but instead of downloading the installer, it will open a web page where Cryptnos can be downloaded.